Try to use token directive in tree
The tree is first created with div tags holding the token's string
representation. This ensures that the nodes are created with the
right size. Afterwards the content from the div is replaced
with the token directive. At the moment for no apparent reason the
last token is bound to the directive.