Trying to make depTree more flexible, but problems
Passing tokens to the dependencyTree is absolutely imperative to make
this reusable.
Thinking of a diff plugin this would mean passing different tokens,
that still work on the selector functions of state - different trees
would highlight the same word.

But something is going wrong here - when I use an isolate scope the
directive fails.
The tree seems to have some strange troubles with not always
registering changes in the state model anyway.

I will try to fix this and see if this problem was also the cause for
the failing isolated scope.