Use keys filter in nestedMenuCollection
We are accessing `all[label]` here twice - trying to solve this with a
variable set through ng-init doesn't work, as ng-init will also resolve
after this tag has been processed (at least that's what I think). If we
want to access this variable from inside the tag - to use it in other
attributes - it won't be set in time and we have to deal with undefined
values.