Change keyCapture directive
It's not placed on the body anymore - this way we have control over
when we want to use it and when not - but more importantly its
initialization is deferred a little, thus we gain access to the
configuration file.

The directive places its event listeners still on the body though.