HACK - make main-body unscrollable
The tree canvas keeps overflowing the main-body and adds an unnecessary
scrollbar to it.
Generally we want scroll functionality in the main-body so we will have
to investigate this further. In a treebanking only environment we won't
have a need for such a scrollbar though - for now we can just disable
scrollability safely and no one will notice...