Deprecate participle as part of speech
The morph retriever returns verb now for all participles. We remove it
as a distinct part of speech category in our morph conf files.

Breaks backwards compatibility. We will have to fix the old treebank
data with this. The introduction of the new schema might be a good
opportunity to do this.