New rules syntax in gr_attributes.json
This is a non-working commit, but done on purpose, as only a single
attribute (Number) is changed. It shall be used for referencing the new
syntax.

Attributes take an array of rule object. Such a rule can have two
properties: "if" and "unless". Both properties have to be objects again.

These objects take other attribute names as keys and a string value or
an array of string values.