Gui prototype of selection buttons
This is an alternative to the mega control for selection and display
of unused tokens. The advantage is that it shows all information
at a glance and is not so complex. The disadvantage is that it
probably needs more screen estate and doesn't scale to a larger
number of possible selections.