Introduce globalStore
This should be a place where retrievers (and plugins) can share data
with each other, that doesn't suit the tokens model of the general
system.

The first use case will be to save constituents. We could save them
sentence wise - but it leads to flexibility and sustainability issues,
as wider code changes would need to be taken into account to do this
every time we introduce a new concept.