Fix ellipsis problem in TreebankPersister
As uncovered by @balmas we were destroying our idMappings during the
handling of new artificialTokens. This was a refactoring gone wrong,
where an older code piece - not needed anymore - was still lingering
around.

We also declare the sentence unchanged after we updated it - this option
was formerly not used at all. If a sentence received changes in an
annotation session, it's changed property stayed true. This meant that
everytime another - unrelated - change was saved in the document, this
sentence would also be re-updated again.

Taking into account that our idMappings were destroyed at this point,
the nasty bug of losing ellided nodes surfaced.