Reactivate `newer` and fix problems with util
I was recently experiencing problems with the `ngtemplates` grunt task,
which aborted the server tasks, because `arethusa.util` does not come
with any templates whatsoever.

I formerly thought this was limited to my desktop machine - something
weird going on with the caching of `grunt-newer`, but as @balmas also
experienced this, we finally have to take measures to prevent this
altogether.

The fix is simple: We just don't add an `ngtemplates:arethusaUtil` task.