prototype hack for #580
This is a complete hack and not done right at all, but serves to provide a
demo of the type of decorator functionality described in #580. Clearly the
setting of the decorator value should be done in the retriever and not the
directive, I was just lazy and didn't feel like parsing the XML.  Probably
this would also be handled by a separate decorator directive and not tacked on
to token, etc.  So this commit should never be merged as-is, just don't want
to lose the work.