add a relocateHandler service and directive
the relocateHandler service is to allow switching between the
different deployments of arethusa (e.g. a main and a staging deployment).
would be a bit nicer if this worked liked the exitHandler as an icon in
the nav bar but I wanted to allow the possibility of supporting multiple
locations and doing a dropdown on the click on the icon was more than
I wanted to deal with for now. So the directive is just a basic list.