show overflowing items
Previously, when a tooltip appeared near the bottom of the screen it
would get cut off, making it impossible for the user to interact with
it. To prevent this from happening, this commit removes
`overflow: hidden` from `body` and `html` in the CSS.

Making this change uncovered a bug in the `fullHeight.js` directive: it
was ignoring padding. This commit also fixes this bug.

These changes together mean that there is no scroll bar until a tooltip
overflows, and when that happens, a scrollbar appears and the user can
interact with it.

(Also included, a tweak to the Docker Compose config to make developing
easier)