Fix bug in forms sorted by frequency
The `morphLocalStorage` module provides functionality to sort forms
based on the frequency they are chosen by a user. Due to a bug in
`toSortedArray`, however, the forms were never being sorted. The result
was that the form that was used first would always appear first in the
list, the one that was used second would be second, etc.

This PR fixes the issue so that forms are sorted based on frequency.