morphology - additional fields
Add an "additionalFields" flag to morphology, similar to the "gloss"
flag. When this is set to true, the following fields will be added to
the morphology tab (and saved in the XML):

* alternateGloss
* semanticRole
* include
* notes

Caveats: this commit doesn't add any tests. There are not tests for
"gloss" and it would be a significant amount of work to update the way
the tests are written to test these fields. The text for these fields is
only in English. It's not the only text that's missing for some of the
languages, but it makes the application slightly less international.