give tree-settings div the correct height
There is an issue with the `dep_tree_no_selector.html` template where
the tree overflows the bounding box. (See
https://github.com/alpheios-project/arethusa/issues/820).

This is caused by the `.tree-settings` div technically having a height
of `0` because its children are all floats. Since it doesn't have any
height, but its children have a height, there is some weird interaction
in a nested svg element that has its own height set to `100%`.

By changing margin to padding and setting overflow to auto, the
`.tree-settings` div gets the correct height which fixes the problem.