display tree if comments not configured
If the comments plugin is not configured but the treebank XML has a
config type that requires comments, then the tree should still be able
to be displayed.

Before this commit, the tree was not able to be displayed because of an
undefined `retriever`. This commit wraps the function call in an if
statement to prevent a crash.